[GH-3] feat: Devise setup with Google sign in (#21)
When a user navigates to the welcome screen, they are now prompted to sign in. When a user signs in, they are permitted to navigate to the welcome screen. Followed the guides provided on omniauth-google-oauth2 repo (https://github.com/zquestz/omniauth-google-oauth2\#devise) and devise wiki (https://github.com/heartcombo/devise/wiki/OmniAuth:-Overview).

Co-authored-by: Royce Threadgill <royce@gnar.dog>