Scope assets resources to gnar (#108)
* Scope assets resources to gnar

When attempting to add the endpoint to display a form to create a new `asset`, a `404 Not Found` was returned and `AssetsController#new` was never hit. Upon inspection of the routes, Propshaft was determined to be the culprit (see the first line item).

```
    Prefix Verb   URI Pattern                Controller#Action
                  /assets                    Propshaft::Server
    assets GET    /assets(.:format)          assets#index
           POST   /assets(.:format)          assets#create
 new_asset GET    /assets/new(.:format)      assets#new
edit_asset GET    /assets/:id/edit(.:format) assets#edit
     asset GET    /assets/:id(.:format)      assets#show
           PATCH  /assets/:id(.:format)      assets#update
           PUT    /assets/:id(.:format)      assets#update
           DELETE /assets/:id(.:format)      assets#destroy
      root GET    /
```

In short, the `/assets` endpoint defined by Propshaft was interfering with the `assets` resources we were attempting to define. Instead of changing the model name and dealing with the ripple effect of doing so, the `assets` resources are now nested under `gnar` which will allow us to keep the data model the same (and maintain the integrity of the G.E.A.R acronym). For example, the route for `Assets#index` will be generated as `/gnar/assets` instead of `/assets`.

* Prefer to namespace assets under gnar

This commit prefers namespacing routes instead of scoping. This decision requires nesting the models, controllers, and views for `assets` under a `Gnar` modile. This is requires a broad code change up front but improves organization.